optimize sprite rendering + make engine dampening areas more visible [...]
I'm too :effort: to split this apart now, but this moves screen coordinate
math into the vertex shader and converts send_sprites from hash to array
operation, making it a good bit faster.

There's also a bugfix in here: For sprites to render correctly, they must
be rendered in the right order. Currently i achieve this by sorting by
texture.

Additionally it darkens colors on various sprites below the floor or above
the ceiling, as well as applies lower alpha to the engine flame sprites.
This makes it clear when the engines aren't effective.

Oh, i also made the engine trails a little wobbly, because.